Organize includes and CMakeLists
