+++ sync-todo/v2/client/cpp/screens/error-modal.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <string>
+
 #include "ftxui/component/captured_mouse.hpp"  // for ftxui
 #include "ftxui/component/component.hpp"       // for Input, Renderer, Vertical
 #include "ftxui/component/screen_interactive.hpp"  // for Component, ScreenInteractive
 #include "ftxui/dom/elements.hpp"  // for text, hbox, separator, Element, operator|, vbox, border
-#include "../app-state.hpp"
 
+#include "../data/app-state.hpp"
 #include "display-screen.hpp"
 
 class ErrorModal {
